Implemented _.filter function
Declared finalArray variable that will be returned at the end. If an
element of the input array satisfies the test then it will be pushed
to my finalArray variable. A for loop was used to iterate over each
array value in the input array and if it satisfies the test function
then it is pushed to the finalArray.