Implemented _.reject function
Used a for loop to iterate over each element in the collection array.
Then used both _.indexOf and _.filter functions to check if element
that is being iterated over is located in the array that would be
returned if the ._filter function was being used. An array is returned
with the values that are not present in the array that is returned by
the result of calling _.filter