Implemented _.uniq function
Start off by creating an object to hold each element that occurs in the
array and then stores the amount of times that that element occurs as
the value.

Declare a finalArray variable at the beginning, which is just an empty
array. We just push each key of the object into the finalArray, but
instead of having the key as a string it is pushed as of type number.