Fix _.reduce function, so that it takes objects as collection value as well
In the body of function, first check if the collection is an array or
object. Use for in loop for object and for loop for arrays. If
accumulator is undefined at the beginning use the first element
of the array or the first value of the initial key in the object.